fastapi-demo/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy[asyncio]==2.0.29
aiosqlite==0.19.0
asyncpg==0.29.0
python-dotenv==1.0.0
strawberry-graphql==0.227.2

